+++ internal/graph/graph.go
@@ -31,6 +31,7 @@ import (
 
 var (
 	errArgKind = errors.New("constructor arguments must be pointers")
+	errRetNode = errors.New("node already exist for the constructor")
 
 	_typeOfError = reflect.TypeOf((*error)(nil)).Elem()
 )
@@ -102,6 +103,10 @@ func (g *Graph) InsertConstructor(ctor interface{}) error {
 		objTypes[i] = ctype.Out(i)
 	}
 
+	if err := g.ValidateReturnTypes(ctype); err != nil {
+		return err
+	}
+
 	nodes := make([]node, count, count)
 	for i := 0; i < count; i++ {
 		nodes[i] = node{
@@ -140,6 +145,24 @@ func (g *Graph) InsertConstructor(ctor interface{}) error {
 		return errors.Wrapf(cycleErr, "unable to Provide %v", objTypes)
 	}
 
+	return nil
+}
+
+// ValidateReturnTypes validates if ctor's return type is already insterted in the graph
+func (g *Graph) ValidateReturnTypes(ctype reflect.Type) error {
+	g.RLock()
+	defer g.RUnlock()
+	objMap := make(map[reflect.Type]bool, ctype.NumOut())
+	for i := 0; i < ctype.NumOut(); i++ {
+		objType := ctype.Out(i)
+		if _, ok := g.nodes[objType]; ok {
+			return errors.Wrapf(errRetNode, "ctor: %v, object type: %v", ctype, ctype.Out(i))
+		}
+		if objMap[objType] {
+			return errors.Wrapf(errRetNode, "ctor: %v, object type: %v", ctype, ctype.Out(i))
+		}
+		objMap[objType] = true
+	}
 	return nil
 }