+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove `*All` funcs to reduce API surface area. Available methods on
 container are `Provide`, `Resolve` and `Invoke`
 - Providing constructors with common returned types results in an error.
+- Update `Provide` and `Resolve` methods to accept variadic arguments.
 
 ## v0.3 (2 May 2017)